Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -8,7 +8,7 @@ int main() {
     int a,b,c;
     scanf("%d %d %d ", &a,&b,&c);
     int sum=a+b+c;
-    int avg= sum/3;
-    printf("Average: %d",avg);
+    float avg= sum/3;
+    printf("Average: %f",avg);
     return 0;
 }
